Refactor appointment booking with multi-table validations and unified common response wrapper

+++ src/main/java/com/stackly1/hospitalmanagementssystem/repository/Doctorrepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.stackly1.hospitalmanagementssystem.entity.Doctor;
 
-public interface Doctorrepository extends JpaRepository<Doctor, Integer> {
+public interface Doctorrepository extends JpaRepository<Doctor, Long> {
 
 }
